requirements..txt
streamlit
opencv-python-headless
numpy
pytesseract
pdf2image
python-pptx
Pillow